Fix HwSwitchEnsemble to not clean up partially created state
Summary:
The HwSwitchEnsemble/BcmSwitchEnsemble classes tried to access their
member programmedState_ in the dtor to cleanup, but if setupEnsemble()
had never been called (e.g., through programming error or exception,
e.g., if the bcm_config file was incorrectly specified or some other
internal invariant was tripped (as was the case when I ran into this),
then the state was never properly populated, and during the dtor various
asan violations and SEGV conditions would arrise.

Added a flag to verify that initialization was done and the dtor checks
the flags before accessing/cleaning up the underlying state.

Also added a test to verify that this all works correctly.  Without the
fix, the test fails.

Also removed the gtest external_dep from the existing packet_util_tests
cpp_unittest because it's apparently no longer compatible with mode/opt
and not needed.

Differential Revision: D18304072

fbshipit-source-id: 63cdab4a96e77df49f40b70e86360d16b77fad6c