Start using Port.profileID to program port resource
Summary:
Start using Port.profileID to look for the correct PortProfileConfig, which contains the modulation, fec information, so we don't have to make assumption in agent code to use the speed decide what modulation to use.

Right now to use the new config to program the PortResource is based on whether the profileID is default or not. If it's default, that means we haven't switched to use the new config yet. It'll fall back to use the old way to program port resource.

Reviewed By: aeckert, birdsoup

Differential Revision: D17814475

fbshipit-source-id: 881972a88a88d64f688821b948f7e5209dd0266b