Enable creating port group from new config
Summary:
The current portgroup logic is based on the theory of each port group is always have 4 logical ports. But some of the new platform demands more flexible port group configuration, which can go up to 8 ports, or probably less, 2 ports.
The new platform config will have a new field in the platform port entry. So we can let each platform to decide its own size of the port group.

Reviewed By: aeckert

Differential Revision: D18072585

fbshipit-source-id: 0e934b932ade90230dbd0289499ab9ae5ea30e19