APIs to get interface MAC from vlan, interface
Summary:
While sending packets in tests we should use
the interface MAC corresponding to the VLAN we are
tagging the packet with. However in nearly all tests
we used platform MAC. This did not matter as long
as the platform and interface MACs were the same,
which is the case in bcm_tests. In sai_tests however
Platform and interface MACs are different, so this
errors shows up. Remember that the platform mac
is just the fallback for the case where interface MAC
is not provided. It's interface MAC that actually gets
programmed in ASIC.

TODO: change the packet send call sites to use interface
MAC instead of platform MAC

#acceept2ship

Reviewed By: sas

Differential Revision: D18666328

fbshipit-source-id: 197a5db05fe7cb011e5117fa994ab1c471f52950