Use interface mac when sending out packets in jumbo frame, l4 port tests
Summary:
As explained in D18666328
"While sending packets in tests we should use
the interface MAC corresponding to the VLAN we are
tagging the packet with. However in nearly all tests
we used platform MAC. This did not matter as long
as the platform and interface MACs were the same,
which is the case in bcm_tests. In sai_tests however
Platform and interface MACs are different, so this
errors shows up. Remember that the platform mac
is just the fallback for the case where interface MAC
is not provided. It's interface MAC that actually gets
programmed in ASIC."

Update jumbo frame and l4 port blackholing tests. More changes
coming in stacked diffs

Reviewed By: sas

Differential Revision: D18666946

fbshipit-source-id: 6c2d553067bfc449888b04dcd4fcbf15831d29ef