Use interface MAC when sending traffic in load balancer tests
Summary:
As explained in D18666328
"While sending packets in tests we should use
the interface MAC corresponding to the VLAN we are
tagging the packet with. However in nearly all tests
we used platform MAC. This did not matter as long
as the platform and interface MACs were the same,
which is the case in bcm_tests. In sai_tests however
Platform and interface MACs are different, so this
errors shows up. Remember that the platform mac
is just the fallback for the case where iinterface MAC
is not provided. Its interface MAC that actually gets
programmed in ASIC"

Update load balancer tests to use interface MAC, this completes
all sendPacketSwitched call sites in Hw tests. Next target is
sendPacketOutOfPort in Hw tests

Differential Revision: D18671076

fbshipit-source-id: 7eb98eda4b583acf21db1173779b84c316805ae8