Use intf mac in InDiscardCounterTest
Summary:
As explained in D18666328
"While sending packets in tests we should use
the interface MAC corresponding to the VLAN we are
tagging the packet with. However in nearly all tests
we used platform MAC. This did not matter as long
as the platform and interface MACs were the same,
which is the case in bcm_tests. In sai_tests however
Platform and interface MACs are different, so this
errors shows up. Remember that the platform mac
is just the fallback for the case where iinterface MAC
is not provided. Its interface MAC that actually gets
programmed in ASIC"

Differential Revision: D18671190

fbshipit-source-id: 4f55c41b5b84d7babf02939bb5e7bb674a408a6a