thrift streaming server support
Summary:
To expose SAI debug shells, we will use thrift streaming (rocket)
This requires a small change on the server side to add a routing service
to the thrift server. I believe that this has no effect on non-stream
APIs, but put it behind an init boolean so we could easily add a GFLAG to
turn it off on non-SAI platforms.

Differential Revision: D18314228

fbshipit-source-id: feb070177cac3574c7e187edde88b66a5f91508c