library to run a python interpreter
Summary:
This library will be wrapped by a thrift diag shell higher in the
stack, so it is primarily concerned with only the "run a python repl"
aspect of things. Since we will end up with different REPLs, this also
represents a proto-interface. I suspect the exact standard stream redirections
will need to be configurable so I went ahead and made that an interface of
the repl.

A few important notes, mostly to do with deficiencies in python:

AFAIK, there is no way to exit PyMain() without crashing the parent program,
so we just keep it running forever. On destruction, leak the thread, assuming
that there is no long-running use case for destroying and re-creating the
thrift handler (it's not so bad to leak one or two threads during a short
lived test, for example)

The alternative, PyRun_InteractiveLoop, seems to be completely busted, with
regard to redirection: https://bugs.python.org/issue14916 (unfixed bug with
patch from community member provided in 2013, last pinged in 2015)

If we want to make this perfect (rather than "working" which is the state in
this diff), we can:
* use lower level PyRun_X primitives to implement the loop
* get the patch prescribed in that bug patched in
* figure out how to exit from PyMain safely (lol)

Differential Revision: D18314226

fbshipit-source-id: 655268b351abc5934e89243f4e22ff10c221e032