Diag Shell -- handles thrift<-->terminal
Summary:
This library takes over some or all of standard streams with a PTY
in order to facilitate the use of a repl over thrift. The PTY behaves like
an "interactive pipe" so we can think of it as a pipe with special behavior
around buffering and special characters.

The basic pattern is:
1. receive a thrift stream publisher in ctor
2. create a PTY (master end is for us, slave end is for repl)
3. point the standard streams the repl expects to use at the slave end
4. run an infinite loop reading the master end and publishing it to the stream
5. if we get a message over thrift, write it to the master end
    typically, this will result in triggering 4. when the repl writes its output
    in response to the new input on the slave end

Differential Revision: D18314225

fbshipit-source-id: fa777d0adc9f16202cbfec762e8fbd29fecd8358