Start thrift handler in Sai tests when --setup_thrift is passed in
Summary:
For now, I am using the same handler as for the main
wedge_agent, even though we don't have a SwSwitch. So only part
of the API is serviceable in Sai tests. In bcm_tests we specialized
then handler and svc interface to only include the APIs needed for
bcm shell. We may yet do that here, but the first target here is
to get shells working

Reviewed By: sas

Differential Revision: D18601392

fbshipit-source-id: 75df571184fd93270c17165ed5eb7eee25adc0f3