Pull out PlatformMode to a separate file
Summary:
There are places where mode is useful without needing the
full ProductInfo, so abstract it out into its own header file.
A further problem with ProductInfo is that it has both oss and facebook
counterparts, so depending on it means you have to fork your dependent
target to oss and facebook too.

Differential Revision: D18722720

fbshipit-source-id: 679a324104a74a5183d84717544af42702efcdd9