Copp tests use TCP packets instead of UDP
Summary:
We have been using UDP packets in all our dataplane
tests. This did not matter as none of our ASIC programming distinguishes
b/w TCP and UDP. This changes with SAI BRCM, there when we install a BGP
trap, it sets up the ACL to match
- IP to me, l4{src, dst} port 179, TCP to hipri queue
This is a more precise definition of BGP packets than the ACLs we use,
which do not qualify on ip proto type. Thus use TCP packets in COPP tests

TODO: Consider changing all tests to use TCP packets since TCP is
90+% of our traffic.

Differential Revision: D18674530

fbshipit-source-id: 6810497a68272a36ccad5bf44f1a9de240b3e3cc