Use the same thrift port number in agent, hw tests
Summary:
When building thrift support for bcm_tests, we picked
a different port number on a whim. This did not matter as we
use bcm_test_shell to interact with bcm_tests and thus could
put a different default port for interacting with bcm_tests.
For sai tests we plan to use the same client for tests and
agent. So unitfy thrift port

Reviewed By: sas

Differential Revision: D18735080

fbshipit-source-id: d0fb8dc33d3f874af5dedb2d7cf9b7db32e55e94