Fix race in HwMacLearning test callback observer
Summary:
Earlier, we sent a packet and then started the observer for learing updates.
However, the callback may be trigerred before the observer starts and thus
might be lost.

This patch fixes it by starting the observer before sending packet.
Specifically, the observer is moved inside the test class itself and started as
part of the test Setup(). This will be particular useful for subsequent tests as
the observer would catch updates callback triggerred even before
setup()/verify()/setupPostWarmboot()/verifyPostWarmboot() is run.

Differential Revision: D18747439

fbshipit-source-id: 6a208dcf6e7a351e686cfada897317f959feb202