Add `RoutingInformationBase::getVrfList()`
Summary:
This is used to get access to the list of existing VRFs inside of the RIB. We
need to add accessor functions like these because the RIB data cannot be
directly exposed to the outside (in particular, because of the use of a
folly::Synchronized object). The first use for this will be in the next commit,
which adds FIB synchronization post-warmboot.

Differential Revision: D18890226

fbshipit-source-id: 493638f0fabe24c4f482db041975037bfc247914