Populate LookupClassUpdater local cache on first notify
Summary:
LookupclassUpdater would get a new neighbor entry with classID already assigned in two cases:
(A) after warmboot: if a neighbor entry had classID associated prior to warmboot.
(B) as a side-effect of classID associated by LookupClassUpdater itself.

In the former case, we should populate the LookupClassUpdater local cache, in latter, we must ignore.

This diff handles it by moving the logic when first state notification is
received by LookupClassUpdater. This helps in subsequent diff which refCnts the mac to classID assignment - with current logic, refCnt would go wrong when LookupClassUpdater receives classID in scenario (B).

Differential Revision: D18876141

fbshipit-source-id: a182905d618f36c5a8034b2662c5396edd0f1587