destroy NeighborUpdaterImpl on NeighborCache thread
Summary:
This moves the destruction of the neighbor caches on to the
neighbor cache thread. Previously, we would waitForUpdates in here
which helps a bit with the race, but it doesn't actually solve it. By
actually destroying on the cache thread we make sure that there are no
active NeighborCacheEntry::process() calls going on and since all data
access is expected to happen on the cache thread it is in line with
the threading model we want.

This should fix S192505.

Reviewed By: boryas

Differential Revision: D19043970

fbshipit-source-id: 122ff845dadb2507437380080c0e332ee4991454