Test values of recently added PortSpeed enums
Summary:
This unit test is trying to test the numeric values of the `PortSpeed` enum. A
couple of values were not tested so this commit adds them.
We can also use a `switch` statement to enforce complete coverage of the
existing enum values by the compiler. Missing checks now result in a build
error like this:
```
[sas@devvm42083 ~/fbsource/fbcode] buck test mode/opt fboss/agent/... -- Thrift.assertPortSpeeds
fboss/agent/test/ThriftTest.cpp:96:13: error: enumeration values 'TWOHUNDREDG' and 'FOURHUNDREDG' not handled in switch [-Werror,-Wswitch]
    switch (key) {
            ^
1 error generated.
```

Differential Revision: D19126835

fbshipit-source-id: 1e7c882e16ffa39858b7d4700678fbba9ce6c493