Support add port queue for newly added port
Summary:
D15787932 starts supporting creating a new port in the software layer. In ThriftConfigApplier::updatePorts, if the cfg port doesn't exist in the port table, we create a new software Port object, which doesn't have any PortQueue.
This will break the logic of ThriftConfigApplier::updatePortQueues(), which expects a non-empty original port queue list because it will always generate the same number of portQueues as original port queues. If the config port queue exist, we'll update the original port queue. Otherwise, we'll create a default port queue, so the hardware can reset such queue to its default value.
```
  // Process all supplied queues
  // We retrieve the current port queue values from hardware
  // if there is a config present for any of these queues, we update the
  // PortQueue according to this
  // Otherwise we reset it to the default values for this queue type
```

However, the to-be created port doesn't have port queue list yet. Besides, ApplyThriftConfig will only be executed once.Therefore, we have to make updatePortQueues work for such case. To do so, I've changed updatePortQueues() to pass in 2 more extra parameters, which are:
- int expectedNumQueues. The expected number of queues we want for a port or cpu.
- cfg::StreamType streamType. The expected stream types of such queue list.

Besides, I also made all the Platforms to implement `getDefaultNumPortQueues()` so that we can use this value to create the expected number of queues for either already created port or to-be created port.

Differential Revision: D18571392

fbshipit-source-id: 426fba320c7d43a646d691545675560c591588c3