load default virtual router from sdk
Summary:
This is a temporary solution to query the default vrf from Switch attribute.

In an ideal scenario, the SDK will provide support for get_object_keys which will be used to load the default virtual router.

When the platform does not support get_object_keys, use the switch attribute to query the default router.

If the SDK did not create a default vrf, create one.

Reviewed By: boryas

Differential Revision: D19157367

fbshipit-source-id: 0d4e439d0a75b5bba3baecf9d3ad8ca742efd59b