make getStatus aware of boot type
Summary:
We talked about this a while back, but on cold boots,
external clients should probably be waiting until the config is
applied, not just that the sdk is initialized.

This makes the agent getStatus call a bit smarter. New logic is that
we will return ALIVE after fully configured on cold boots, but after
initialized on warm boots.

We may want to just always wait until configured. Advertising we are
ready earlier could matter if we were rolling out some required
configuration change across warm boot and openr/bgp expected that
configuration to be present before starting up. However, this is
sufficient for now.

Reviewed By: boryas

Differential Revision: D19276183

fbshipit-source-id: 29c3c52a831042f6c4eed9c3bb346af3da76fdc9