dump previous and new state on failed state application
Summary:
We hit a one off problem that was triggered by a bad state
update in canary. Have a rough idea of problem area, but kind of hit
end of debugging I could do. In order to have an easier way to
reproduce crashes we see in prod, I added logic that dumps the old/new
state for the state update that led to a crash so it is easier to
understand the change that triggered a crash.

It would also be good to try to write some sort of "replay" tool that
takes in these two states, cold boots in to old state and then applies
the new state. Could help with reproing some of these problems or
easily adding test cases from real prod problems.

Reviewed By: phshaikh

Differential Revision: D19373394

fbshipit-source-id: 8e89b1464cbe5b0bd548a1fb38d2feb2e3634901