open source facebook/BcmSwitchSettings.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmSwitchSettings.cpp to open source.
This is achieved as follows:
- hg rm fbcode/fboss/agent/hw/bcm/oss/BcmSwitchSettings.cpp
- Copy over contents of bcm/facebook/BcmSwitchSettings.cpp to bcm/BcmSwitchSettings.cpp
- hg rm fbcode/fboss/agent/hw/bcm/facebook/BcmSwitchSettings.cpp
- Edit CMakeLists to remove oss/BcmSwitchSettings.cpp
- Edit bcm/TARGETS to remove facebook/BcmSwitchSettings.cpp and oss/BcmSwitchSettings.cpp

Differential Revision: D19419439

fbshipit-source-id: bf523d213a924a0b5a0b323aca6e49932c4a754f