open source facebook/BcmMacTable.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmMacTable.cpp to open source.
This is achieved as follows:
- hg rm fbcode/fboss/agent/hw/bcm/oss/BcmMacTable.cpp
- Copy over contents of bcm/facebook/BcmMacTable.cpp to bcm/BcmTrunk.cpp
- hg rm fbcode/fboss/agent/hw/bcm/facebook/BcmMacTable.cpp
- Edit CMakeLists to remove oss/BcmMacTable.cpp
- Edit bcm/TARGETS to remove facebook/BcmMacTable.cpp and oss/BcmTrunk.cpp

Differential Revision: D19419001

fbshipit-source-id: b462c06061e0a1d9ef90d702acabcad0c13ad89c