open source facebook/BcmTrunk.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmTrunk.cpp to open source.
This is achieved as follows:
 - hg rm fbcode/fboss/agent/hw/bcm/oss/BcmTrunk.cpp
 - Copy over contents of bcm/facebook/BcmTrunk.cpp to bcm/BcmTrunk.cpp
 - hg rm fbcode/fboss/agent/hw/bcm/facebook/BcmTrunk.cpp
 - Edit CMakeLists to remove oss/BcmTrunk.cpp
 - Edit bcm/TARGETS to remove facebook/BcmTrunk.cpp and oss/BcmTrunk.cpp

Differential Revision: D19418241

fbshipit-source-id: 4001fe12019796cc10556a8bf795e7abc45cdd0b