open source facebook/BcmEgress.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmEgress.cpp to open source.
This is achieved as follows:
- hg rm fbcode/fboss/agent/hw/bcm/oss/BcmEgress.cpp
- Copy over contents of bcm/facebook/BcmEgress.cpp to bcm/BcmEgress.cpp
- hg rm fbcode/fboss/agent/hw/bcm/facebook/BcmEgress.cpp
- Edit CMakeLists to remove oss/BcmEgress.cpp
- Edit bcm/TARGETS to remove facebook/BcmEgress.cpp and oss/BcmEgress.cpp

Differential Revision: D19424286

fbshipit-source-id: 37d12ada021da5f2a83e8adba2e99bf3a76b90c4