open source facebook/BcmHost.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmHost.cpp to open source.
This is achieved as follows:
- hg rm fbcode/fboss/agent/hw/bcm/oss/BcmHost.cpp
- Copy over contents of bcm/facebook/BcmHost.cpp to bcm/BcmHost.cpp
- hg rm fbcode/fboss/agent/hw/bcm/facebook/BcmHost.cpp
- Edit CMakeLists to remove oss/BcmHost.cpp
- Edit bcm/TARGETS to remove facebook/BcmHost.cpp and oss/BcmHost.cpp

Differential Revision: D19424288

fbshipit-source-id: 296465b91f6983227bfe9123bedb738142efc271