open source facebook/BcmSwitchEventCallback.cpp
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmSwitchEventCallback.cpp to open source.
This is achieved as follows:
- move facebook/BcmSwitchEventCallback.cpp to open source
- open source must include soc/opensoc.h, the diff uses the fact that open
  source does not define BCM_VER_MAJOR to conditionally include right headers

Differential Revision: D19424287

fbshipit-source-id: 1750067a5731538e4346505dda17b26347e18e51