open source facebook/BcmQosMap*
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA. All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

Towards that goal, this diff moves facebook/BcmQosMap functionality to open
source.

As part of it, we need to (and now can) open source BcmQosMapEntry as well.

Differential Revision: D19424285

fbshipit-source-id: 45b04e02f40f1396d140d75b0526311d92c1ef4c