use new config structures when programming port in agent
Summary:
NOTE: This diff requires a full rollout of the configuration changes and changes to coop in D17814475, D17563627, etc. This should not be landed until the new configuration structs have been rolled out.

This diff only uses the new configuration to program the external phy. The old config is currently still used to program the internal phy.

Differential Revision: D18175885

fbshipit-source-id: 3b0635cf0b2598b7f64119ef2337e81ed0cbf78a