open source BcmSwitch.cpp and dependencies
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

This patch open sources and BcmSwitch.cpp dependencies
This is somewhat tricky as diag shell is not open sourced.

Differential Revision: D19462189

fbshipit-source-id: 68b713b616111ac9eecf837f014fce11c3102bcb