open source BcmAPI, BcmFacebookAPI and dependencies
Summary:
With the landing of D16401189, FBOSS OSS no longer depends on OpenNSL,
but instead links with OpenNSA.  All the BCM API is available in OpenNSA.

Thus, start moving agent/hw/bcm/facebook/* to open source.

This patch open sources BcmAPI, BcmFacebookAPI and dependencies.
This is somewhat tricky as diag shell is not open sourced.

Differential Revision: D19461875

fbshipit-source-id: ec61937c7017378ea80eb9f84af1ea6a010f7437