Bubble up hwFeaturesDesired to init platform functions
Summary:
As titled. We need this information when creating a
HwSwitch, which happens in initPlatform functions. So bubble
this information up.
This change is a noop for wedge_agent since it always creates
a HwSwtich with default HwSwitch features (linkscan, packet rx)
enabled. But this diff will help clean up hw test code.

Stacked diff to make use of this functionality in hw tests

Differential Revision: D19443537

fbshipit-source-id: 16edd318f6e0f6d762b93d459571b04f0fe3fce3