Only expose applyInitialConfigAndBringUpPorts from link state toggler
Summary:
As titled, config application is not really complete till
we have applied the config and brought up ports - we deliberately
keep ports down by setting preemphasis to 0 and lb mode to NONE
in applyInitialConfig and then only later (in bringUpPorts)
reflect the desired loopback mode on ports enabled in config.

For external clients - config application is really only done
once we have reflected the status of ports as reflected
by initial config.

Next diff will rename this function to just applyInitialConfig
but I wanted to keep this as separate so as to not pollute with
code for changing names at the callsites.

Differential Revision: D19414220

fbshipit-source-id: a4dd14ba05374b977cd10427985a39b8ac975664