Implement set attribute in FakeSaiHostIf and fix test
Summary:
We forgot to add this set attribute implementation, which
meant that in Fake hostif trap attributes were only being set
during create. There was a test that exercised this code path,
but did not uncover the bug as
- The test only checked the value set on sai object and not what
the value was on the underlying sai object. Fix this now
- SaiObject set attribute is not checking return value of api.setAttribute.
This will be fixed in subsequent diff

Reviewed By: boryas

Differential Revision: D19558529

fbshipit-source-id: ae8319e05c9f80f8388e8ce84e29c9e48d4d21bb