Start refactoring OSS CMake
Summary:
The buck TARGETS for fboss creates several libraries that get reused  across
agent/test code. However, OSS CMake often puts all sources into single binary.

This has several downsides:
- We may have to repeat same sources in different binaries.
- Hard to find source files that are missing from CMake.
- Hard to maintain.
- As we start building SAI in OSS, the code common to BCM and SAI is best built
  as a library. The buck TARGETS already do that.

It would be easier for CMake to follow lib hierarchy similar to TARGETS.

This is a step in that direction:

In general, libraries and binaries in fboss/foo/bar would be built by a cmake
file FooBar.cmake.
Once this transition is complete, we could add butterfly rule to TARGETS, so
anyone modifying the TARGETS could do similar (and trivial) change to CMake as
well.

There are 276 binaries and libraries under fboss (outside of facebook), that
would need to be converted if we go with this approach (so not-trivial), but I
think the one time effort will be worth it. Would like to get thoughts too.

Differential Revision: D19575588

fbshipit-source-id: f14775e4be197702d7ff106943c67305bfbac812