Add module loopback support for miniphoton pims.
Summary: Since Miniphoton PIMs have optic module on board, we are not able to put eloop to replace the optic module to test the healthness of other components on that path of the port. Thus, for debugging, we have this extra byte to set loopback within the module. And this is the software support of that feature.

Differential Revision: D19238504

fbshipit-source-id: 5f5c4788dbecd28a211f0a5043285372235e7fe2