get base physical lane when adding ports
Summary:
Currently BcmPortResourceBuilder only gets lanes when removing the ports. The problem is that, when going from 1 port to 2 ports, the only port that's checked for the base lanes is the higher port, so the base lane is higher than it should be.

To fix this, I've made the code also check the added ports to see if any of them have a lower base lane.

Differential Revision: D18944346

fbshipit-source-id: 242529006a91591d2e0c47405dcf2809fdc892ed