Make DataPlanePhyChip use map
Summary:
In the config, we use a list to store all the DataPlanePhyChip for a platform, which can be less efficient for our code to quickly access a chip by name.
So I introduced
```
const std::map<std::string, phy::DataPlanePhyChip>& Platform::getDataPlanePhyChips()
```
and used a map to maintain all the chips in PlatformMapping.

Reviewed By: birdsoup

Differential Revision: D19532861

fbshipit-source-id: 1b3dd6004848e8c5bd414f074e6ba90a97bd687c