Tmp fix for xphy reprogram on warmboot from medium/interfaceMode change
Summary: A recent change added medium and interfaceMode to the platform config for YAMP (P125979578). Right now we don't support reading interfaceMode/medium from the baldeagle code, so the configs always mismatch and cause the port to be reprogrammed on a warmboot. This change is a temporary fix to unbreak future pushes once we've re-landed D19532871 until we can actually compare the medium/interfaceMode.

Differential Revision: D19746150

fbshipit-source-id: 2412e87f69a04cbc6a2e155873654558f19b41ae