Handle MAC Move in LookupClassUpdater
Summary:
We got burned by this during queue-per-host rollout, because consider following sequence:

- Agent looks up eth0’s MAC, sets locally administered bit, say MAC M0.
- Agent sends packet with M0 as source MAC on downlink. M0 is learned on cpu0.
- Consider a case when such a packet is looped back on a downlink port D.
- This results in MAC Move, two callbacks: DELETE M0 from cpu0, ADD M0 for D.
- From LookupClassUpdater’s perspective, this will be processNeighborChanged<folly::MacAddress>, where implementation is missing (I had a TODO!): https://fburl.com/diffusion/2fh346rp

If CPU continues to send packets, the MAC will ping pong between cpu port and downlink port.

The CPU probably sends packets frequently enough so that this MAC is not aged often (aging = 300s).
However, consider a case when the MAC ages out, and the MAC is learned on downlink port.

Because the downlink port would have non-zero list of classIDs, it would trigger a call to removeClassIDForPortAndMac, which will find that the refCnt is 0 (the MAC was never added). This could be rare enough explaining why we did not see this issue often. But, this also means that agent can hit this issue at any time during operation.

Fix: handle MAC Move.

We should consider disabling MAC Move for this special local MAC to avoid this ping pong.

We should investigate (and if possible prevent) why we are getting such packets looped back.

Differential Revision: D19827929

fbshipit-source-id: ea16d24962a186f807b65f6b861c0fedb30ae7bf