Fixed a bug that caused SwSwitch::getPortStatus to return incomplete results
Summary:
Due to a recent deprecation of the front panel in the config, `WedgePort::frontPanel_` is not populated for TH3 platforms. This caused `WedgePort::getChannel` to fail which caused the port status to have a TransceiverIdx with an empty channels list.

If the channels list is empty, `fboss int sum` wont be fully populated because the CLI assumes ports without channels are not interesting.

Differential Revision: D19727748

fbshipit-source-id: af13f7a3fba96c2b815ba6b01f1d2bb26a966e16