Sai Api - add api to enable querying a subset of stats
Summary:
Not all stats for a given type would work on all chips.
So provide a api which allows for taking a vector of stat_ids
(which can be composed at runtime). This way we can now parameterize
based on the ASIC type.

nit: We should probably change the interface of getStats to take
a range/pair of iterators. Will try that in a subsequent diff.

Reviewed By: phshaikh

Differential Revision: D19956747

fbshipit-source-id: 7228bace6a66fac079811e3dc59617e1b265440d