SFINAE for overload resolution in getAttribute to support rvalue as well as lvalue tuple argument
Summary:
As titled.

Defining a lvalue of tuple and and calling  getAttribute to fetch tuple of attributes seemed a bit unnecessary, as tuple is passed as non-const reference but new tuple is returned. Usage seemed a bit off the beaten path to me. Would be much more convenient to pass prvalue as an argument to this function, but existing parameter was lvalue reference and so can't take prvalue.

Providing universal reference in getAttribute while getting Tuple of attributes. This should permit all arguments of all value categories.

Also putting some useful SFINAE in culling candidates based on arguments of getAttribute to make it more readable. Importantly, std::optional has implicit constructor for prvalue, so compiler complained about ambiguity, fixed it with SFINAE

Reviewed By: boryas

Differential Revision: D19966756

fbshipit-source-id: 6256002f4f5461eb1f7e33c01f14941fb85c332b