Gate Label switch action tests on MPLS support
Summary:
As titled, these tests fail on TD2, but
we haven't disabled them. Instead, we have marked
them as known bad. This is sub optimal, as now when
we are trying to move TD2 to start running ALPM mode
we see what look like fresh failures (since the name
changes from wedge_lpm to wedge_alpm), but are actually
already broken.

Use this knob so oncall is not bugged with tasks for
these tests

Differential Revision: D20064265

fbshipit-source-id: cd883a2a798e8dc0f35afa76f405634487f0772f