Make PlatformMapping related functions const
Summary:
Since the following function:
```
std::vector<PortID> BcmTestPlatform::getAllPortsinGroup(PortID portID) const
```
is a const function and needs to use `numPortsPerTranceiver_` to generate the port id list in the group.

This might be a bit difficult to support the future platforms, which might have different size of port group.

Since we're switching to use PlatformMapping to get the platform config, which will also contain the controlling port and subsidiary ports relationship. And these get functions can be const too if we just use the PlatformMapping to get the config.

Differential Revision: D19920713

fbshipit-source-id: 78ff6653f8f9a983a85b1d314f0ace99ec32632e