Enable sai hw tests to support only diag shell interface.
Summary:
Today when we run thrift in hw_tests on SAI it supports the entire SaiCtrl interface.

This includes the generic FbossCtrl interface, which depends on a SwSwitch being present. Since there is no SwSwitch in hw tests, this gets us into trouble.

We add a new SaiTestCtrl interface that follows the approach of bcm tests where we abstracted out a BcmTestCtrl interface that was supported by bcm_tests.

Differential Revision: D19908946

fbshipit-source-id: 894c9bb2c2316e2c2ac8645a3b141c16de294758