Skip programming subnet route for /32 and /128 interfaces
Summary:
SaiSwitch currently has a bug in the case where we configure an
interface address with a subnet of size 1 (/32 for v4 and /128 for v6).
Programming the interface creates a ToMe route, but then we overwrite it
with a subnet route and break the ToMe functionality. We considered fixing
it directly in the RIB by having that be the source of the ToMe routes (rather
than creating them in an adhoc fashion in SaiRouterInterfaceManager and BcmIntf
but that ran into significant issues in BcmHost and the "use host table for /32
routes" optimization, so we have opted for this band-aid here for expediency.

Reviewed By: phshaikh

Differential Revision: D20044540

fbshipit-source-id: 4bd2fea79269ed186478269efdda3913cae307b4