Clean up warm boot helper class hierarchy
Summary:
Get rid of gratuitous inheritance in BcmWarmBootHelper.
We only ever do Disc backed warm boots, at some point we had
a idea of doing memory backed warm boots for FakeSDK and this
hierarchy was born. We never went that way, and even if
we choose to - we would just use tmp dir like so
https://fburl.com/diffusion/kvdx8iko

This refactor will make it easier for me to abstract away
common code and share it for bcm and sai warm boots.

Reviewed By: phshaikh

Differential Revision: D20126457

fbshipit-source-id: a811645cde30f5ef382bd05625becd559269a1ef