Add SaiInSegEntryManager.cpp to oss build
Summary:
D19576628 added SaiInSegEntryManager.cpp.

There was no build failure despite the missing file as sai_switch is built with
unresolved-symbols=ignore-all, and that setting is never disabled (it should be
disabled for sai_test that links with sai_switch lib). That fix will follow
soon, but for now add SaiInSegEntryManager.cpp to OSS build.

Reviewed By: phshaikh

Differential Revision: D20148768

fbshipit-source-id: fe09f5440ccc30cf5249c9f0800f935f4ad66bb2