global lock shared by all SaiApis
Summary:
The SAI spec does not provide any guarantees about the thread safety
of the SAI adapter. In fact, we know that BCM SAI does not provide thread
safe access to its API. As a result, we must be pessimistic and lock s.t.
only one thread is ever inside a SAI API.

Reviewed By: phshaikh

Differential Revision: D20229250

fbshipit-source-id: d4329fca9b311eb6c67cfe986616c5599f559283