use fmt::formatter to generically print SaiAttribute
Summary:
For error/informational messages, it would be really useful to be
able to print SaiAttributes (and tuples of them!). This diff shows off how
we might be able to do it, using the popular fmt library. I would have
preferred to use folly::Format or std::format but the former is actually
noticeably worse than fmt and the latter doesn't quite exist yet (c++20).

One area I am actively looking for ideas/suggestions is how we customize
the string for each SaiAttribute (see RouteApi.h/cpp). I think what I've
done is OK, but maybe someone can think of a better way. Other approaches
I considered: some kind of compile time string that could be a non-type
template parameter of SaiAttribute (couldn't figure out how to do it) and
changing the declarations of SaiAttributes from aliases to trivial inheritance
that adds a str() method.

Next steps are printing SaiObject in store/ as well as printing the full
contents of each operation in error/debug logs.

Reviewed By: phshaikh

Differential Revision: D20171960

fbshipit-source-id: 9bd695ffe4e859acb17e9fadf94966af42220127