Wedge agent LED thrift control
Summary:
Enables the control of LEDs through a thrift call, to aid in the process of diagnostics/repair for the DC people, discussion can be found [here](https://fb.workplace.com/groups/fim.development/permalink/2268208663479530/).

FIM team included this now as a possible feature enhancement for the 'test driven cabling' initiative they are working on with ENS/GDCC/Netplatform.

Differential Revision: D19973334

fbshipit-source-id: 0b4c865cfe50567921b4e1dd8331bc39c35e6951