Alter FBOSS cli utility to set LED values externally via thrift
Summary:
Enables the control of LEDs through a thrift call, to aid in the process of diagnostics/repair for the DC people, discussion can be found [here](https://fb.workplace.com/groups/fim.development/permalink/2268208663479530/).

FIM team included this now as a possible feature enhancement for the 'test driven cabling' initiative they are working on with ENS/GDCC/Netplatform.

Differential Revision: D20159887

fbshipit-source-id: ee1038ba69eab2d253caf2f0f60fc7448b106a7c