Add a knob to convey whether a asic needs object key caching
Summary:
For BRCM SAI implementations, we only have partial
get_object_keys support (enough to be able to build up
cold boot state with ports and such). So to work around
that, we write object keys to disc on exit and retrieve them
during warm boot initialization.
Abstracted out a BroadcomAsic class to put in common apis
for these ASICs.
I thought of putting in a feature OBJECT_KEYS_GET and then
turning it off of BRCM ASIC, however as noted above, we
have implemented partial support for this to get us to
a sane cold boot state (ports show up post init), so we
can't just have the feature be off.

Reviewed By: boryas

Differential Revision: D20368654

fbshipit-source-id: cd12573d460c5b3d2ad801e52d39f51d38b39d24