fboss: skip detach if the unit is not attached
Summary:
wedge_agent always creates BCM unit and attach to it. During deinitialization,
we blindly call into detach. When adding tests for platform only code, calling
into detach causes crash because BCM unit is not initialized in this case.

The diff adds checks to skip detach if the unit is not attached.

Differential Revision: D20331358

fbshipit-source-id: 151ee163be4dfb4394026de973d69d3bfe8146f1