Fix galaxy front panel port logic
Summary:
As D20375947, we also need to correct how to use frontPanel_ after we switched to use PlatformMapping to initialize a platform. Otherwise, the frontPanel_ will be nullopt and all the port will be treated as BackPlanePort in Galaxy.

Besides, since the transceiver id should be fixed after we create WedgePort object. We can just use `private std::optional<TransceiverID> transceiverID_;` to represent whether this PlatformPort has a transceiver. And we can assign the value to transceiverID_ in the constructor and make getTransceiverID() to return transceiverID_ directly.

Reviewed By: aeckert

Differential Revision: D20380943

fbshipit-source-id: 8bb88bc17b9a9c7e09adf4030c579063138baf54