simplify/fix NeighborCache shutdown
Summary:
We are still seeing issues with shutdown of NeighborCache
since switching the module to all run on its own thread. Initially
this was noticed in S192505 & presumably fixed w/ D19043970. However,
the destruction itself actually asynchronously schedules destroying
the underlying entries. This leads to the following interleaving:

NeighborCacheEVB
- start destroying NeighborUpdater/NeighborCache
- schedule destroying all NeighborCacheEntries
- finish destroying NeighborUpdater/NeighborCache
- NeighborCacheEntry(s) race between destruction/waking up and segfaulting

This diff removes the apis to explicitly clear all the entries. Since
the destructor and all other work is run on the NeighborCache evb,
there is no need for a complicated shutdown order any more. Instead,
we can destroy everything normally synchronously.

Reviewed By: boryas

Differential Revision: D20413860

fbshipit-source-id: 3a9868736d25a390f0958e842fdd09b5054d3ab0