publisher object traits
Summary:
a useful trait for objects which are monitored or subscribed for by subscriber. Will be using it going forward.

an example is neighbor entry, which is 'publisher' for next hop.

another example is bridge port, which is 'publisher' for fdb entry.

A publisher is an object which publishes some attributes. These published attributes are subscribed by other subscriber objects.

IsObjectPublisher -> a trait to identify if its state change such as create/delete must be published
ObjectPublisherAttributes -> a metafunction which returns what are published attributes or what attributes of this publisher governs life of subscriber

Reviewed By: boryas

Differential Revision: D20150051

fbshipit-source-id: ab33b67c2fd97b3aab959e5ec38c4906ee52bc6e