fix LookupClassUpdaterTests post D20483648
Summary:
Waiting for all the threads after certain actions is now even
more complex since changing neighbor cache to a proxying
model.

D20483648 changed to properly proxy the lookup class call to fix a
race condition where it was accessing NeighborCache synchronously from
the update thread. However, this breaks some test assumptions.

Instead of waiting for the state update to complete and knowing
we have gone from update observer -> NeighborUpdater -> NeighborCache
-> scheduling the resulting update inline, we now have to do more of a
dance between the neighborcache thread and the update thread.

All we know when the initial update completes is that we have
scheduled something on the neighborcache thread. Then we wait for
neighbor cache thread and we should have scheduled something back on
the update thread. Wait for update thread here and we should see the
result show up in switch state.

This fixes how we wait to handle this new world order.

Reviewed By: shri-khare

Differential Revision: D20525970

fbshipit-source-id: 46ea287a44b5f41b28a4363853bc36489b830b6f