stop proxying NeighborUpdater port(Up|Down) through background evb
Summary:
No idea how this code lasted so long. We purged the
background evb from this module ages ago.

The code as written will proxy to bg thread and then immediately proxy
to the neighbor cache thread, which is unnecessary. This diff changes
the logic to just directly proxy the computation to the neighbor cache
thread.

Differential Revision: D20527673

fbshipit-source-id: 54ef056363715eb2e3e9671a9b54f73b2e6f22eb