sai object event subscriber
Summary:
Boiler plate templates for
1) subscriber interface
2) single event subscriber
3) aggregate event subscriber

Subscriber interface: This is used by publisher, it provides minimum things such as
1) notify subscriber object of create or delete of publisher
2) save opaque subscription token by publisher

This also provides a utility methods to subclasses to
1) store weak reference to publisher
2) maintain publisher attributes (that govern subscriber)
3) accessors to both

single event subscriber: This is an implementation of interface. It is a facade to more complex aggregate subscriber, which is used by publisher to dispatch notifications as above

aggregate subscriber:  a collection or mixin of  one or more single event subscribers. a facade to more complex subscriber implementation which can be arbitrarily complex. A subscriber implementation can subscribe to more than one publishers using aggregate subscriber, leading to N-to-N relationship between publishers and subscribers. aggregate subscribers main jobs involve
1) aggregating events from all subscribers
2) informing subscriber implementation about these events if
2.1) all publishers are ready
2.2) any one publisher has died

Additionally aggregate subscriber gives "boiler plate" methods to set and reset subscriber implementation's "SAI" object.

A subclass of this class is a  subscriber implementation which must follow some "protocol" as expected by this  class
1) aggregate subscriber  informs "subscriber impl object" if all publishers are available
2) "subscriber impl object" must prepare create attributes & adapter host keys depending on which subscriber's SAI object is being implemented
3) "subscriber impl object" then invokes "setObject" this constructs a SAI object
4) aggregate subscriber  informs "subscriber impl object" if any publisher is dying
5) "subscriber impl object" can update attributes and either update "SAI object" or destroy it.

Reviewed By: boryas

Differential Revision: D20150049

fbshipit-source-id: 9c529d5d64064affcc120b2fb5295c84be90b3dc