fix pointless use of folly::demangle in printing fboss strong types
Summary:
folly::demangle has compiler-specific behavior and as a result, can
cause unit tests to fail and unwanted output depending on the compiler. It
turns out, I overlooked the fact that we have access to the new type name as
part of the macro's parameters, so it is much easier to just inject it with
macro quoting. I tried a few ways to not have to use fmt:: and '#' but couldn't
make it work. This seems good enough :)

Reviewed By: bkoray

Differential Revision: D20497754

fbshipit-source-id: 8e8d4c2a0d569c13d45a84769c152e266e182840